pyproject.toml
[tool.poetry]
name = "mortier"
version = "0.1.0"
description = ""
authors = ["Epsilon"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
svgwrite = "^1.4.3"
numpy = "^2.3.1"
pillow = "^11.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
